win32: some compilation fixes here and there.

+++ src/half.h
@@ -28,9 +28,9 @@ public:
     /* Constructors. Always inline so that the code can work in registers
      * instead of calling routines with the hidden "this" parameter. */
     inline half() { }
-    inline half(int f) { *this = makefast(f); }
+    inline half(int f) { *this = makefast((float)f); }
     inline half(float f) { *this = makefast(f); }
-    inline half(double f) { *this = makefast(f); }
+    inline half(double f) { *this = makefast((float)f); }
 
     inline int is_nan() const
     {
@@ -53,9 +53,9 @@ public:
     }
 
     /* Cast to other types -- always inline, see constructors */
-    inline half &operator =(int f) { return *this = makefast(f); }
+    inline half &operator =(int f) { return *this = makefast((float)f); }
     inline half &operator =(float f) { return *this = makefast(f); }
-    inline half &operator =(double f) { return *this = makefast(f); }
+    inline half &operator =(double f) { return *this = makefast((float)f); }
     inline operator int() const { return (int)tofloat(*this); }
     inline operator float() const { return tofloat(*this); }
